Refactor class telegram into own module
