+++ application.py
@@ -1,14 +1,17 @@
-import requests
 import configparser
 from ogame import OGame
-from ogame.constants import destination, coordinates, ships, mission, speed, buildings, status
+from core.telegram import Telegram
 
 # Read config file
 config = configparser.ConfigParser()
 config.read('config.cfg')
 login = config['Login']
-telegram = config['Telegram']
+telegramKey = config['Telegram']
 
+# Inits
+print('### Init service start')
+telegram = Telegram(telegramKey.get('Token'), telegramKey.get('Chat'))
+print('Telegram service is started')
 
 # block: user infos
 print('### User informations')
@@ -25,30 +28,21 @@ print('')
 print('### Some informations about the server')
 server = empire.server()
 print('The server version is', server.version)                    
-print('The sppe of the server is', server.Speed.universe)             
-print('The Spped of the fleet is', server.Speed.fleet)                
+print('The speed of the server is', server.Speed.universe)
+print('The speed of the fleet is', server.Speed.fleet)
+
+# block: probe
+print('### Run probe')
+# Send a Spyprobe to a random Player
 
 
-class Telegram:
-    def __init__(self):
-        self.TOKEN = telegram.get('Token')
-        self.CHAT = telegram.get('Chat')
-        self.URL = "https://api.telegram.org/bot{}".format(self.TOKEN)
 
-    def get_url(self, url):
-        response = requests.get(url)
-        content = response.content.decode("utf8")
-        return content
 
-    def send_message(self, text):
-        url = self.URL + "/sendMessage?text={}&chat_id={}".format(text, self.CHAT)
-        self.get_url(url)
-
-#Telegram informations
-Telegram().send_message('---')
-Telegram().send_message('Ogame bot ist gestartet. User {0} in the uni {1} was successfull logged in.'.format(login.get('Username'), login.get('Uni')))
-if empire.characterclass():
-  Telegram().send_message('The User has the class {0}'.format(empire.characterclass()))
+#telegram informations
+telegram.send_message('---')
+telegram.send_message('Ogame bot ist gestartet. User {0} in the uni {1} was successfull logged in.'.format(login.get('Username'), login.get('Uni')))
+if not empire.characterclass():
+  telegram.send_message('The User has the class {0}'.format(empire.characterclass()))
 if empire.planet_ids():
   for planetID, planetName in zip(empire.planet_ids(), empire.planet_names()):
-    Telegram().send_message('The name of the planet is [{1}] with id [{0}]'.format(planetID, planetName))
+    telegram.send_message('The name of the planet is [{1}] with id [{0}]'.format(planetID, planetName))
